habitat-practice/python/requirements.txt
alembic==1.8.0
attrs==21.4.0
Automat==20.2.0
beautifulsoup4==4.11.1
blis==0.7.8
catalogue==2.0.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
constantly==15.1.0
crochet==2.0.0
cryptography==37.0.4
cssselect==1.1.0
cycler==0.11.0
cymem==2.0.6
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
filelock==3.7.1
Flask==2.1.2
Flask-Cors==3.0.10
fonttools==4.34.4
greenlet==1.1.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.12.0
incremental==21.3.0
install==1.3.5
itemadapter==0.6.0
itemloaders==1.0.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
kiwisolver==1.4.3
langcodes==3.3.0
lxml==4.9.1
Mako==1.2.1
MarkupSafe==2.1.1
matplotlib==3.5.2
MechanicalSoup==1.1.0
murmurhash==1.0.7
nltk==3.7
numpy==1.23.0
packaging==21.3
pandas==1.4.3
parsel==1.6.0
pathy==0.6.2
Pillow==9.2.0
preshed==3.0.6
Protego==0.2.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.1
PyDispatcher==2.0.5
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==5.4.1
queuelib==1.6.2
regex==2022.7.9
requests==2.28.1
requests-file==1.5.1
scikit-learn==1.1.1
scipy==1.8.1
Scrapy==2.6.1
service-identity==21.1.0
six==1.16.0
smart-open==5.2.1
soupsieve==2.3.2.post1
spacy==3.4.0
spacy-legacy==3.0.9
spacy-loggers==1.0.2
SQLAlchemy==1.4.39
srsly==2.4.3
thinc==8.1.0
threadpoolctl==3.1.0
tldextract==3.3.0
tqdm==4.64.0
Twisted==22.4.0
typer==0.4.2
typing_extensions==4.3.0
urllib3==1.26.9
w3lib==1.22.0
wasabi==0.9.1
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
zope.interface==5.4.0
